# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Olopress API",
        "version": "1.0"
    },
    "paths": {
...
        "/auth/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate the user",
                "operationId": "auth_token",
                "requestBody": {
                    "description": "Authenticate",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResult"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/roles/findAll": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Find all roles",
                "operationId": "roles_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAll"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/roles/findOneById/{roleId}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Find item data",
                "operationId": "roles_findOneById",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RolesFindOneById"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/roles/update/{roleId}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update role",
                "operationId": "roles_update",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update role",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthRequest": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AuthResult": {
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserObject"
                    },
                    "expiresAt": {
                        "description": "Expiration date of token",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "UserObject": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CommonFindAll": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "RoleSave": {
                "properties": {
                    "roleId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "roleKey": {
                        "type": "string"
                    },
                    "roleName": {
                        "type": "string"
                    },
                    "roleLevel": {
                        "type": "number"
                    },
                    "rolePermissions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "permId": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "RolesFindOneById": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/RolesFindOneByIdObject"
                    }
                },
                "type": "object"
            },
            "RolesFindOneByIdObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "roleName": {
                        "type": "string"
                    },
                    "roleKey": {
                        "type": "string"
                    },
                    "roleLevel": {
                        "type": "string"
                    },
                    "rolePermissions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "permId": {
                                    "type": "string"
                                },
                                "moduleName": {
                                    "type": "string"
                                },
                                "method": {
                                    "type": "string"
                                },
                                "route": {
                                    "type": "string"
                                },
                                "action": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}